Refactors site summary report generation logic
Reimplements site summary report logic by moving it from a partially converted VB.NET file to a new, dedicated C# helper class. This provides a cleaner implementation that integrates directly with FMS domain entities for improved maintainability.

Updates display text for abandoned/inactive status tracker comments.

Relates to #991
